requirements.txt
Django>=5.2.1
djangorestframework>=3.16.0
django-cors-headers>=4.7.0
transformers>=4.38.2
torch>=2.2.1
pytesseract>=0.3.13
Pillow>=11.2.1
python-dotenv>=1.0.0
requests>=2.32.3
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pydantic==2.5.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
alembic==1.12.1
psycopg2-binary==2.9.9 